requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# LLM & Embeddings
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5

# Vector Store
chromadb==0.4.22

# Sparse Retrieval
rank-bm25==0.2.2

# Document Processing
python-docx==1.1.0
pypdf==4.0.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
python-multipart==0.0.9

# Firebase (Optional)
firebase-admin==6.4.0

# Development
pytest==8.0.0
pytest-asyncio
httpx==0.26.0
